Moved mockMobileContent to NewestMobile folder
