+++ src/components/NewestMobile/mockMobileContent.tsx
